Player elmaslara geldiğinde elmaslar teker teker gelmesi sağlandı. Kazıcıların özellikleri datalaştırıldı.

+++ Assets/Scripts/Controllers/HostageMinnerController/HostageMinnerAIController.cs
@@ -34,7 +34,7 @@ namespace Controllers.HostageController
             _coalTargetList = MiningDistricSignalable.Instance.onCoalsList?.Invoke();
         }
 
-        public void GetAgentSpeed(int Speed)
+        public void GetAgentSpeed(float Speed)
         {
             navMeshAgent.speed = Speed;
         }
